requirements.txt
flask>=3.0.0
pymongo>=4.6.1
python-dotenv>=1.1.0
requests>=2.31.0
dnspython>=2.4.2
gunicorn>=21.2.0
fastmcp>=2.3.0
arize-phoenix-otel>=0.6.0
openinference-instrumentation-mcp>=0.1.0
opentelemetry-exporter-otlp>=1.20.0